Various fixes and improvements suggested by SonarQube.
Fixes include:
- don't refer to subclass from parent class initialization.
- cast ints to long before assignment to avoid overflow
- Iterator.next() method should throw NoSuchElementException
- if getter is synchronized, setter should also be (and vice versa)
- if base method is synchronized, override should be as well
- unary minus on compareTo() result can cause problems if result is MIN_VALUE
- group regex to make | alternatives clearer
- don't use several overrides for compareTo() (see RelationInfo)

+++ engine/src/main/java/nl/inl/blacklab/codec/SegmentRelationInfo.java
@@ -173,11 +173,11 @@ public class SegmentRelationInfo implements AutoCloseable {
             try {
                 // Determine where the relations for this doc start
                 assert docsOffset >= 0 : "negative offset in docs file";
-                docs().seek(docsOffset + docId * Long.BYTES);
+                docs().seek(docsOffset + (long) docId * Long.BYTES);
                 long relationsOffset = _docs.readLong();
                 assert relationsOffset >= 0 : "negative offset in relations file";
                 // Find the attribute set offset for this relation
-                relations().seek(relationsOffset + relationId * Long.BYTES);
+                relations().seek(relationsOffset + (long) relationId * Long.BYTES);
                 long attrSetOffset = _relations.readLong();
                 // Find the attribute set
                 assert attrSetOffset >= 0 : "negative offset in attrSet file";
